Fix error in conversion to make parallelization possible
The timesync synchronization function assumed that only one file in a
folder would be converted at a time, and consequently, temporary
conversion output was saved to a file with a set filename. This
temporary file's name is now based on the files it is converting, so
that multiple files can be synchronized at once.